UPDATE enable editor in CreateStory and UpdateStory
